修复生成seatmap时的问题
